requirements.txt
Flask
Flask-Limiter
gunicorn
selenium
python-telegram-bot
requests

